Rename the /components/admin directory to /components/users
