+++ src/components/users/CreateMultipleUsers.tsx
@@ -2,7 +2,7 @@ import type { NextPage } from 'next'
 import { useToast, Stack, Button } from '@chakra-ui/react'
 import { parse, type ParseResult } from 'papaparse'
 import { trpc } from '~/utils/trpc'
-import DataTable from '~/components/admin/DataTable'
+import DataTable from '~/components/users/DataTable'
 import { add } from '~/store/admin/dashboard'
 import { useDispatch, useSelector } from 'react-redux'
 import type { RootState } from '~/store/store'
